ci_f_ncp {confintr}	R Documentation

CI for the Non-Centrality Parameter of the F Distribution

Description

Based on the inversion principle, parametric CIs for the non-centrality parameter (NCP) Delta of the F distribution are calculated. To keep the input interface simple, we do not provide bootstrap CIs here.

Usage

ci_f_ncp(x, df1 = NULL, df2 = NULL, probs = c(0.025, 0.975))

Arguments

x
The result of stats::lm() or the F test statistic.

df1
The numerator df. Only used if x is a test statistic.

df2
The denominator df. Only used if x is a test statistic.

probs
Lower and upper probabilities, by default c(0.025, 0.975).

Details

A positive lower (1 - α) \cdot 100\%-confidence limit for the NCP goes hand-in-hand with a significant F test at level α. According to stats::pf(), the results might be unreliable for very large F values.

Value

An object of class "cint", see ci_mean() for details.

References

Smithson, M. (2003). Confidence intervals. Series: Quantitative Applications in the Social Sciences. New York, NY: Sage Publications.

See Also

ci_rsquared()

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
ci_f_ncp(fit)
ci_f_ncp(fit, probs = c(0.05, 1))

[Package confintr version 1.0.2 Index]