ci_mad {confintr}	R Documentation

CI for the MAD

Description

This function calculates bootstrap CIs (default: "bca") for the population median absolute deviation (MAD), see stats::mad() for more information.

Usage

ci_mad(
  x,
  probs = c(0.025, 0.975),
  constant = 1.4826,
  type = "bootstrap",
  boot_type = c("bca", "perc", "norm", "basic"),
  R = 9999L,
  seed = NULL,
  ...
)

Arguments

x
A numeric vector.

probs
Lower and upper probabilities, by default c(0.025, 0.975).

constant
Scaling factor applied. The default (1.4826) ensures that the MAD equals the standard deviation for a theoretical normal distribution.

type
Type of CI. Currently not used as the only type is "bootstrap".

boot_type
Type of bootstrap CI c("bca", "perc", "norm", "basic").

R
The number of bootstrap resamples. Only used for type = "bootstrap".

seed
An integer random seed. Only used for type = "bootstrap".

...
Further arguments passed to boot::boot().

Value

An object of class "cint", see ci_mean() for details.

Examples

x <- rnorm(100)
ci_mad(x, R = 999)  # Use larger R

[Package confintr version 1.0.2 Index]