Confidence Interval for the Population Variance

This function calculates confidence intervals for the population variance. By default, classic confidence intervals are calculated based on the chi-squared distribution, assuming normal distribution (see Smithson). Bootstrap confidence intervals are also available and are recommended for the non-normal case as the chi-squared confidence intervals are sensitive to deviations from normality.

ci_var(
  x,
  probs = c(0.025, 0.975),
  type = c("chi-squared", "bootstrap"),
  boot_type = c("bca", "perc", "stud", "norm", "basic"),
  R = 9999,
  seed = NULL,
  ...
)

Arguments

x

A numeric vector.

probs

Error probabilites. The default c(0.025, 0.975) gives a symmetric 95% confidence interval.

type

Type of confidence interval. One of "chi-squared" (default) or "bootstrap".

boot_type

Type of bootstrap confidence interval ("bca", "perc", "stud", "norm", "basic"). Only used for type = "bootstrap".

R

The number of bootstrap resamples. Only used for type = "bootstrap".

seed

An integer random seed. Only used for type = "bootstrap".

...

Further arguments passed to boot::boot.

Value

A list with class cint containing these components:

parameter: The parameter in question.

interval: The confidence interval for the parameter.

estimate: The estimate for the parameter.

probs: A vector of error probabilities.

type: The type of the interval.

info: An additional description text for the interval.

Details

Bootstrap confidence intervals are calculated by the package "boot", see references. The default bootstrap type is "bca" (bias-corrected accelerated) as it enjoys the property of being second order accurate as well as transformation respecting (see Efron, p. 188). The "stud" (bootstrap t) bootstrap uses a general formula for the standard error of the sample variance given in Wilks.

References

Smithson, M. (2003). Confidence intervals. Series: Quantitative Applications in the Social Sciences. New York, NY: Sage Publications.

S.S. Wilks (1962), Mathematical Statistics, Wiley & Sons.

Efron, B. and Tibshirani R. J. (1994). An Introduction to the Bootstrap. Chapman & Hall/CRC.

Canty, A and Ripley B. (2019). boot: Bootstrap R (S-Plus) Functions.

See also

Examples

x <- 1:100
ci_var(x)
#> 
#> 	Two-sided 95% chi-squared confidence interval for the population
#> 	variance
#> 
#> Sample estimate: 841.6667 
#> Confidence interval:
#>      2.5%     97.5% 
#>  648.8375 1135.8202 
#> 
ci_var(x, type = "bootstrap", R = 999)
#> 
#> 	Two-sided 95% bootstrap confidence interval for the population variance
#> 	based on 999 bootstrap replications and the bca method
#> 
#> Sample estimate: 841.6667 
#> Confidence interval:
#>     2.5%    97.5% 
#> 716.4407 999.7748 
#>