kurtosis {confintr}	R Documentation

Pearson's Measure of Kurtosis

Description

Defined as the ratio of the 4th central moment and the squared second central moment. Under perfect normality, the kurtosis equals 3. Put differently, we do not show "excess kurtosis" but rather kurtosis.

Usage

kurtosis(z, na.rm = TRUE)

Arguments

z
A numeric vector.

na.rm
Logical flag indicating whether to remove missing values or not. Default is TRUE.

Value

Numeric vector of length 1.

See Also

moment(), skewness()

Examples

kurtosis(1:10)
kurtosis(rnorm(1000))

[Package confintr version 1.0.2 Index]