Confidence Interval for the Non-Centrality Parameter of the F Distribution

Based on the inversion principle, parametric confidence intervals for the non-centrality parameter Delta of the F distribution are calculated. Note that we do not provide bootstrap confidence intervals here to keep the input interface simple. A positive lower (1-alpha)*100%-confidence limit for the ncp goes hand-in-hand with a significant F test at level alpha.

Usage

ci_f_ncp(x, df1 = NULL, df2 = NULL, probs = c(0.025, 0.975))

Arguments

x

The result of lm or the F test statistic.

df1

The numerator degree of freedom, e.g. the number of parameters (including the intercept) of a linear regression. Only used if x is a test statistic.

df2

The denominator degree of freedom, e.g. n - df1 - 1 in a linear regression. Only used if x is a test statistic.

probs

Error probabilites. The default c(0.025, 0.975) gives a symmetric 95% confidence interval.

Value

A list with class cint containing these components:

parameter: The parameter in question.

interval: The confidence interval for the parameter.

estimate: The estimate for the parameter.

probs: A vector of error probabilities.

type: The type of the interval.

info: An additional description text for the interval.

Details

Note that, according to ?pf, the results might be unreliable for very large F values.

References

Smithson, M. (2003). Confidence intervals. Series: Quantitative Applications in the Social Sciences. New York, NY: Sage Publications.

See also

Examples

fit <- lm(Sepal.Length ~ ., data = iris)
ci_f_ncp(fit)
#> 
#> 	Two-sided 95% F confidence interval for the non-centrality parameter of
#> 	the F-distribution
#> 
#> Sample estimate: 980.4737 
#> Confidence interval:
#>      2.5%     97.5% 
#>  704.6598 1200.8313 
#> 
ci_f_ncp(fit, probs = c(0.05, 1))
#> 
#> 	One-sided 95% F confidence interval for the non-centrality parameter of
#> 	the F-distribution
#> 
#> Sample estimate: 980.4737 
#> Confidence interval:
#>       5%     100% 
#> 738.8387      Inf 
#> 
ci_f_ncp(fit, probs = c(0, 0.95))
#> 
#> 	One-sided 95% F confidence interval for the non-centrality parameter of
#> 	the F-distribution
#> 
#> Sample estimate: 980.4737 
#> Confidence interval:
#>      0%     95% 
#>    0.00 1155.07 
#> 
ci_f_ncp(x = 188.251, df1 = 5, df2 = 144)
#> 
#> 	Two-sided 95% F confidence interval for the non-centrality parameter of
#> 	the F-distribution
#> 
#> Sample estimate: 980.474 
#> Confidence interval:
#>     2.5%    97.5% 
#>  704.660 1200.832 
#>